fix(iris): pass long count to Arena.allocateArray for MRT texture array
Arena.allocateArray(ValueLayout, long) requires a long count; passing
the int colorCount directly failed compilation with "cannot find symbol
method allocateArray(AddressLayout, int)". Cast to long.

Native Swift (buildMacNative) compiled fine — the MRT native functions
and Unmanaged<MTLTexture> borrowing are valid.

+++ src/main/java/com/metallum/client/metal/render/MetalCommandEncoder.java
@@ -222,7 +222,7 @@ final class MetalCommandEncoder implements CommandEncoderBackend {
         // The texture pointer array only needs to live for the duration of the
         // native call, so a confined arena closed immediately after is safe.
         try (Arena arena = Arena.ofConfined()) {
-            MemorySegment texArray = arena.allocateArray(ValueLayout.ADDRESS, colorCount);
+            MemorySegment texArray = arena.allocateArray(ValueLayout.ADDRESS, (long) colorCount);
             for (int i = 0; i < colorCount; i++) {
                 MemorySegment handle = colorTextureViews[i] == null
                         ? MemorySegment.NULL : colorTextureViews[i].nativeHandle();
